Remade the process based installer to use the neo4j.bat on windows. Refactured the download helper by splitting the extraction code out of it

+++ Neo4j.Driver/Neo4j.Driver.IntegrationTests/WindowsProcessBasedNeo4jInstaller.cs
@@ -3,7 +3,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Neo4j.Driver.IntegrationTests
@@ -12,15 +11,25 @@ namespace Neo4j.Driver.IntegrationTests
   /// A process based Neo4j installer that uses a background process for 
   /// running the Neo4j server. 
   /// </summary>
-  public class ProcessBasedNeo4jInstaller : INeo4jInstaller
+  public class WindowsProcessBasedNeo4jInstaller : INeo4jInstaller
   {
     private const string eventSourceName = "Neo4jTests";
+
+    private readonly int delayAfterStartingServer;    
+
+    public WindowsProcessBasedNeo4jInstaller(int delayAfterStartingServer = 15000)
+    {
+      this.delayAfterStartingServer = delayAfterStartingServer;
+    }
 
     public DirectoryInfo Neo4jHome { get; private set; }
 
     public void DownloadNeo4j()
     {
-      Neo4jHome = Neo4jDownloadHelper.DownloadNeo4j();
+      Neo4jHome = Neo4jServerFilesDownloadHelper.DownloadNeo4j(
+        Neo4jServerEdition.Enterprise, 
+        Neo4jServerPlatform.Windows, 
+        new ZipNeo4jServerFileExtractor());
 
       UpdateSettings(new Dictionary<string, string> { { "dbms.security.auth_enabled", "false" } });// disable auth
     }
@@ -36,13 +45,11 @@ namespace Neo4j.Driver.IntegrationTests
       try
       {
         ProcessStartInfo processInfo = new ProcessStartInfo();
-        processInfo.FileName = @"java.exe";
+        processInfo.FileName = Path.Combine(Neo4jHome.FullName, @"bin\neo4j.bat");
         processInfo.WorkingDirectory = Neo4jHome.FullName;
-        processInfo.Arguments = string.Format(@"-cp ""{0}lib/*;{0}plugins/*"" -server -Dorg.neo4j.config.file=conf/neo4j.conf -Dlog4j.configuration=file:conf/log4j.properties -Dneo4j.ext.udc.source=zip-powershell -Dorg.neo4j.cluster.logdirectory=data/log -Dorg.neo4j.config.file=conf/neo4j.conf -XX:+UseG1GC -XX:-OmitStackTraceInFastThrow -XX:hashCode=5 -XX:+AlwaysPreTouch -XX:+UnlockExperimentalVMOptions -XX:+TrustFinalNonStaticFields -XX:+DisableExplicitGC -Dunsupported.dbms.udc.source=zip -Dfile.encoding=UTF-8 org.neo4j.server.CommunityEntryPoint", Neo4jHome.FullName);
-        processInfo.UseShellExecute = false;
-        processInfo.CreateNoWindow = true;
-        processInfo.RedirectStandardInput = true;
-        processInfo.RedirectStandardOutput = true;
+        processInfo.Arguments = "console";
+        processInfo.UseShellExecute = true;
+        processInfo.CreateNoWindow = false;
 
         EventLog.WriteEntry(eventSourceName, $"Starting process: {processInfo.FileName} with working directory: {processInfo.WorkingDirectory} and arguments: {processInfo.Arguments}", EventLogEntryType.Information);
 
@@ -54,32 +61,7 @@ namespace Neo4j.Driver.IntegrationTests
         throw;
       }
 
-      Task task = new Task(() =>
-      {
-        StringBuilder logCollectlor = new StringBuilder();
-        while (!startedProcess.HasExited)
-        {
-          string line = startedProcess.StandardOutput.ReadLine();
-          logCollectlor.AppendLine(line);
-        }
-
-        string log = logCollectlor.ToString();
-        if (log.Contains("ERROR"))
-        {
-          EventLog.WriteEntry(eventSourceName, log, EventLogEntryType.Error);
-
-          // This exception is only seen when debugging unittest
-          // This can be helpful debugging if/why the Neo4j server does not start/work
-          throw new InvalidOperationException(log);
-        }
-        else
-        {
-          EventLog.WriteEntry(eventSourceName, log, EventLogEntryType.Information);
-        }
-      });
-      task.Start();
-
-      Task.Delay(15000).Wait();
+      Task.Delay(delayAfterStartingServer).Wait();
     }
 
     public void StopServer()
@@ -89,12 +71,12 @@ namespace Neo4j.Driver.IntegrationTests
       try
       {
         EventLog.WriteEntry(eventSourceName, "Stopping process");
-        startedProcess.StandardInput.AutoFlush = true;
-        startedProcess.StandardInput.Close();
         startedProcess.CloseMainWindow();
-        Task.Delay(10000).Wait();
-        startedProcess.Kill();
-        startedProcess.WaitForExit();
+        if (!startedProcess.WaitForExit(10000))
+        {
+          startedProcess.Kill();
+          startedProcess.WaitForExit();
+        }
         startedProcess = null;
       }
       catch (Exception ex)
